FU-54 refactor: cognito 등록 및 토큰 발행 절차를 successHandler에서 처리하도록 수정
기존 코드에서는 카카오에서 사용자 정보를 가져오는 시점에 cognito 유저 풀에 등록하고 토큰을 발행했음.
하지만 이때 토큰을 발행하게 되면 인증이 끝나고 redirect 시키는 시점에서 accessToken, refreshToken을 쿠키에 세팅하기 위해 토큰을 다시 얻어오는 등 유사한 작업을 반복해야하므로 수행 시점을 리팩토링하게 되었음.
